move manymerge directory to test
